refactor(insights): extract shared Insights_REST_Trait from controllers
