+++ plugins/newspack-plugin/includes/wizards/insights/api/class-gates-rest-controller.php
@@ -26,8 +26,6 @@ namespace Newspack\Insights;
 defined( 'ABSPATH' ) || exit;
 
 use DateTimeImmutable;
-use DateTimeZone;
-use Exception;
 use WP_Error;
 use WP_REST_Controller;
 use WP_REST_Request;
@@ -39,6 +37,7 @@ use WP_REST_Server;
 class Gates_REST_Controller extends WP_REST_Controller {
 
 	use Cached_Controller_Trait;
+	use Insights_REST_Trait;
 
 	/**
 	 * Shared Tab 4–7 namespace.
@@ -114,29 +113,6 @@ class Gates_REST_Controller extends WP_REST_Controller {
 	}
 
 	/**
-	 * Permission check.
-	 *
-	 * This route is intentionally callable via application passwords. The
-	 * BQ-side rate limit is the 10-minute cooldown enforced in
-	 * {@see Cache::refresh()}, not a per-route rate limiter — any caller
-	 * authenticated as a user with `manage_options` (whether via cookie +
-	 * nonce or an application password) can trigger a refresh, and the
-	 * cooldown applies uniformly.
-	 *
-	 * @return bool|WP_Error
-	 */
-	public function permissions_check() {
-		if ( ! current_user_can( 'manage_options' ) ) {
-			return new WP_Error(
-				'newspack_insights_rest_forbidden',
-				__( 'You do not have permission to view Insights data.', 'newspack-plugin' ),
-				[ 'status' => rest_authorization_required_code() ]
-			);
-		}
-		return true;
-	}
-
-	/**
 	 * GET handler.
 	 *
 	 * @param WP_REST_Request $request Request.
@@ -206,59 +182,6 @@ class Gates_REST_Controller extends WP_REST_Controller {
 				return $this->build_response( $metric, $start, $end, $compare_start, $compare_end );
 			}
 		);
-	}
-
-	/**
-	 * Validate and parse the window args. Returns [start, end, compare_start, compare_end] on
-	 * success; WP_Error on validation failure.
-	 *
-	 * @param WP_REST_Request $request Incoming request.
-	 * @return array|WP_Error
-	 */
-	private function parse_window_args( WP_REST_Request $request ) {
-		$tz = $this->site_timezone();
-		try {
-			$start = $this->parse_date( $request->get_param( 'start' ), $tz, false );
-			$end   = $this->parse_date( $request->get_param( 'end' ), $tz, true );
-		} catch ( Exception $e ) {
-			return new WP_Error( 'newspack_insights_invalid_date', $e->getMessage(), [ 'status' => 400 ] );
-		}
-		if ( $start > $end ) {
-			return new WP_Error(
-				'newspack_insights_invalid_window',
-				__( 'Start date must be on or before end date.', 'newspack-plugin' ),
-				[ 'status' => 400 ]
-			);
-		}
-
-		$compare_start_param = $request->get_param( 'compare_start' );
-		$compare_end_param   = $request->get_param( 'compare_end' );
-		$compare_start       = null;
-		$compare_end         = null;
-		if ( $compare_start_param || $compare_end_param ) {
-			if ( ! $compare_start_param || ! $compare_end_param ) {
-				return new WP_Error(
-					'newspack_insights_invalid_comparison',
-					__( 'Both compare_start and compare_end must be provided to enable comparison mode.', 'newspack-plugin' ),
-					[ 'status' => 400 ]
-				);
-			}
-			try {
-				$compare_start = $this->parse_date( $compare_start_param, $tz, false );
-				$compare_end   = $this->parse_date( $compare_end_param, $tz, true );
-			} catch ( Exception $e ) {
-				return new WP_Error( 'newspack_insights_invalid_date', $e->getMessage(), [ 'status' => 400 ] );
-			}
-			if ( $compare_start > $compare_end ) {
-				return new WP_Error(
-					'newspack_insights_invalid_comparison_window',
-					__( 'compare_start must be on or before compare_end.', 'newspack-plugin' ),
-					[ 'status' => 400 ]
-				);
-			}
-		}
-
-		return [ $start, $end, $compare_start, $compare_end ];
 	}
 
 	/**
@@ -389,104 +312,5 @@ class Gates_REST_Controller extends WP_REST_Controller {
 			// Section 3 empty-state totals (NPPD-1694).
 			Gates_Metric::paywall_section_totals( $paywall_direct, $paywall_influenced )
 		);
-	}
-
-	/**
-	 * Args spec.
-	 *
-	 * @return array
-	 */
-	public function get_collection_params() {
-		$base = [
-			'type'              => 'string',
-			'sanitize_callback' => 'sanitize_text_field',
-			'validate_callback' => [ $this, 'validate_date_string' ],
-		];
-		return [
-			'start'         => array_merge(
-				$base,
-				[
-					'description' => __( 'Inclusive window start date (YYYY-MM-DD, site timezone).', 'newspack-plugin' ),
-					'required'    => true,
-				]
-			),
-			'end'           => array_merge(
-				$base,
-				[
-					'description' => __( 'Inclusive window end date (YYYY-MM-DD, site timezone).', 'newspack-plugin' ),
-					'required'    => true,
-				]
-			),
-			'compare_start' => array_merge(
-				$base,
-				[
-					'description' => __( 'Optional comparison window start. Must pair with compare_end.', 'newspack-plugin' ),
-					'required'    => false,
-				]
-			),
-			'compare_end'   => array_merge(
-				$base,
-				[
-					'description' => __( 'Optional comparison window end. Must pair with compare_start.', 'newspack-plugin' ),
-					'required'    => false,
-				]
-			),
-		];
-	}
-
-	/**
-	 * REST validate_callback.
-	 *
-	 * @param mixed $value Value.
-	 * @return bool|WP_Error
-	 */
-	public function validate_date_string( $value ) {
-		if ( ! is_string( $value ) || '' === $value ) {
-			return new WP_Error(
-				'newspack_insights_invalid_date',
-				__( 'Date must be a non-empty YYYY-MM-DD string.', 'newspack-plugin' ),
-				[ 'status' => 400 ]
-			);
-		}
-		$parsed = DateTimeImmutable::createFromFormat( 'Y-m-d', $value, $this->site_timezone() );
-		if ( ! $parsed || $parsed->format( 'Y-m-d' ) !== $value ) {
-			return new WP_Error(
-				'newspack_insights_invalid_date',
-				/* translators: %s: the invalid date string */
-				sprintf( __( 'Invalid date "%s". Expected YYYY-MM-DD.', 'newspack-plugin' ), $value ),
-				[ 'status' => 400 ]
-			);
-		}
-		return true;
-	}
-
-	/**
-	 * Parse a Y-m-d string into a DateTimeImmutable.
-	 *
-	 * @param mixed        $value      Raw value.
-	 * @param DateTimeZone $tz         Timezone.
-	 * @param bool         $end_of_day If true, 23:59:59; else 00:00:00.
-	 * @return DateTimeImmutable
-	 * @throws Exception On parse failure.
-	 */
-	private function parse_date( $value, DateTimeZone $tz, bool $end_of_day ): DateTimeImmutable {
-		if ( ! is_string( $value ) || '' === $value ) {
-			throw new Exception( esc_html__( 'Missing date value.', 'newspack-plugin' ) );
-		}
-		$parsed = DateTimeImmutable::createFromFormat( 'Y-m-d', $value, $tz );
-		if ( ! $parsed || $parsed->format( 'Y-m-d' ) !== $value ) {
-			/* translators: %s: the invalid date string */
-			throw new Exception( esc_html( sprintf( __( 'Invalid date "%s". Expected YYYY-MM-DD.', 'newspack-plugin' ), $value ) ) );
-		}
-		return $end_of_day ? $parsed->setTime( 23, 59, 59 ) : $parsed->setTime( 0, 0, 0 );
-	}
-
-	/**
-	 * Site timezone.
-	 *
-	 * @return DateTimeZone
-	 */
-	private function site_timezone(): DateTimeZone {
-		return wp_timezone();
 	}
 }
